Move common `new_partial()` to domains/service/src/lib.rs
